fetch my store profile
