+++ app/src/main/java/com/alya/ecommerce_serang/data/api/retrofit/ApiService.kt
@@ -63,5 +63,5 @@ interface ApiService {
 
 
     @GET("mystore")
-    fun getStore (): Call<StoreResponse>
+    suspend fun getStore (): Response<StoreResponse>
 }
