+++ app/src/main/java/com/alya/ecommerce_serang/ui/product/ProductViewModel.kt
@@ -4,9 +4,9 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.alya.ecommerce_serang.data.api.dto.Store
 import com.alya.ecommerce_serang.data.api.response.Product
 import com.alya.ecommerce_serang.data.api.response.ReviewsItem
-import com.alya.ecommerce_serang.data.api.response.Store
 import com.alya.ecommerce_serang.data.repository.ProductRepository
 import kotlinx.coroutines.launch
