+++ app/src/main/java/com/alya/ecommerce_serang/ui/profile/mystore/MyStoreActivity.kt
@@ -1,0 +1,67 @@
+package com.alya.ecommerce_serang.ui.profile.mystore
+
+import android.os.Bundle
+import android.widget.Toast
+import androidx.activity.enableEdgeToEdge
+import androidx.activity.viewModels
+import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.viewModels
+import com.alya.ecommerce_serang.data.api.dto.Store
+import com.alya.ecommerce_serang.data.api.dto.UserProfile
+import com.alya.ecommerce_serang.data.api.retrofit.ApiConfig
+import com.alya.ecommerce_serang.data.api.retrofit.ApiService
+import com.alya.ecommerce_serang.data.repository.MyStoreRepository
+import com.alya.ecommerce_serang.data.repository.UserRepository
+import com.alya.ecommerce_serang.databinding.ActivityMyStoreBinding
+import com.alya.ecommerce_serang.ui.profile.ProfileViewModel
+import com.alya.ecommerce_serang.utils.BaseViewModelFactory
+import com.alya.ecommerce_serang.utils.SessionManager
+import com.bumptech.glide.Glide
+import kotlin.getValue
+
+class MyStoreActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityMyStoreBinding
+    private lateinit var apiService: ApiService
+    private lateinit var sessionManager: SessionManager
+
+    private val viewModel: MyStoreViewModel by viewModels {
+        BaseViewModelFactory {
+            val apiService = ApiConfig.getApiService(sessionManager)
+            val myStoreRepository = MyStoreRepository(apiService)
+            MyStoreViewModel(myStoreRepository)
+        }
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMyStoreBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        sessionManager = SessionManager(this)
+        apiService = ApiConfig.getApiService(sessionManager)
+
+        enableEdgeToEdge()
+
+        viewModel.loadMyStore()
+
+        viewModel.myStoreProfile.observe(this){ user ->
+            user?.let { myStoreProfileOverview(it) }
+        }
+
+        viewModel.errorMessage.observe(this) { error ->
+            Toast.makeText(this, error, Toast.LENGTH_SHORT).show()
+        }
+    }
+
+    private fun myStoreProfileOverview(store: Store){
+
+        binding.tvStoreName.setText(store.storeName.toString())
+        binding.tvStoreType.setText(store.storeType.toString())
+
+//        if (store.storeImage != null && user.image is String) {
+//            Glide.with(this)
+//                .load(user.image)
+//                .into(binding.profileImage)
+//        }
+    }
+}
