+++ app/src/main/res/layout/activity_my_store.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.profile.mystore.TokoSayaActivity">
+    tools:context=".ui.profile.mystore.MyStoreActivity">
 
     <include layout="@layout/header" />
